Create README.md
I am Akash Tripathi and I am from India and I  am a student with a vision to become a data analyst or a web developer. Intrested in money making projects💘.......
Also want an opportunity to learn various projects in python or in various other languages. I am a biginner with a aim to make carrier in CS.